requirements.txt
aiogram==3.7.0
python-dotenv==1.0.1
openpyxl~=3.1.5